Add interactive API documentation at /docs/api
resources, so the request schemas come from the actual validation rules
and cannot drift from the code.

- /docs/api      browsable UI with a working "Try It"
- /docs/api.json OpenAPI 3.1 spec, importable into Postman or used to
                 generate a Dart client

All 12 endpoints are documented, each marked public or bearer-only based
on its middleware. The overview covers the token flow, the refresh-on-401
rule, the response envelope and the status codes.

Docs are public so the Flutter and web developers can work against the
deployed API; API_DOCS_ENABLED=false hides them behind a 404.

Two contract fixes found while reviewing the generated spec:
- expires_in was typed string; a Dart client expecting int would fail
- otp/request returned 200 or 201 depending on whether a row was created,
  which is an implementation detail and produced a useless response
  schema. Always 200 now.

+++ app/Services/Auth/TokenService.php
@@ -47,7 +47,9 @@ class TokenService
             'access_token' => $accessToken->plainTextToken,
             'refresh_token' => $plainRefresh,
             'token_type' => 'Bearer',
-            'expires_in' => $ttlMinutes * 60,
+            // Cast so the generated API schema types this as an integer;
+            // a Dart client expecting int would fail on a string.
+            'expires_in' => (int) ($ttlMinutes * 60),
         ];
     }
